% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_kerasjs.R
\name{convert_hdf5model.kerasjs_conversion}
\alias{convert_hdf5model.kerasjs_conversion}
\title{Converts HDF5 Model to KerasJS}
\usage{
\method{convert_hdf5model}{kerasjs_conversion}(model_name = NULL,
  target = "kerasjs", ...)
}
\arguments{
\item{model_name}{The path to the HDF5 exported model, as a string.}

\item{target}{The target path for the converted model.}
}
\description{
Converts HDF5 model into KerasJS format.
}
