% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_interface.R
\name{convert_hdf5model}
\alias{convert_hdf5model}
\title{Converts a HDF5 Model}
\usage{
convert_hdf5model(model_name = NULL, format = c("kerasjs"),
  target = file_path_sans_ext(model_name), ...)
}
\arguments{
\item{model_name}{The path to the HDF5 exported model, as a string.}

\item{format}{The target format for the converted model. Currently only
\code{kerasjs}.}

\item{target}{The target path for the converted model.}
}
\description{
Converts a HDF5 model into other model formats.
}
