% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_interface.R
\name{convert_savedmodel}
\alias{convert_savedmodel}
\title{Converts a SavedModel}
\usage{
convert_savedmodel(model_dir = NULL, format = c("tflite"),
  target = paste("savedmodel", format, sep = "."),
  signature_name = "serving_default", ...)
}
\arguments{
\item{model_dir}{The path to the exported model, as a string.}

\item{format}{The target format for the converted model. Currently only
\code{tflite}.}

\item{target}{The target path for the converted model.}

\item{signature_name}{The named entry point to use in the model for prediction.}

\item{...}{Additional arguments. See \code{?convert_savedmodel.tflite_conversion}
for additional options.}
}
\description{
Converts a TensorFlow SavedModel into other model formats.
}
