% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tflite.R
\name{convert_savedmodel.tflite_conversion}
\alias{convert_savedmodel.tflite_conversion}
\title{Convert to TensorFlow Lite}
\usage{
\method{convert_savedmodel}{tflite_conversion}(model_dir = NULL,
  target = "savedmodel.tflite", signature_name = "serving_default",
  inference_type = "FLOAT", quantized_input_stats = NULL,
  drop_control_dependency = TRUE, ...)
}
\arguments{
\item{model_dir}{The path to the exported model, as a string.}

\item{target}{The target path for the converted model.}

\item{signature_name}{The named entry point to use in the model for prediction.}

\item{quantized_input_stats}{For each member of input_tensors the mean and
std deviation of training data. Only needed \code{inference_type} is
\code{"QUANTIZED_UINT8"}.}

\item{...}{Additional arguments. See \code{?convert_savedmodel.tflite_conversion}
for additional options.}

\item{inference_type:}{Currently must be \code{"FLOAT"} or
\code{"QUANTIZED_UINT8"}.}

\item{drop_control_dependency:}{Drops control dependencies silently. This is
due to tf lite not supporting control dependencies.}
}
\description{
Converts a model to TensorFlow Lite.
}
