% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{kerasjs_convert}
\alias{kerasjs_convert}
\title{Converts HDF5 Model to KerasJS}
\usage{
kerasjs_convert(model_name = NULL, target = file_path_sans_ext(model_name))
}
\arguments{
\item{target}{The target path for the converted model.}

\item{model_dir}{The path to the exported model, as a string.}

\item{...}{Additional arguments, currently not in use.}
}
\description{
Converts HDF5 model into KerasJS format.
}
